requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.15.0
redis==3.5.3
vk-api==11.9.0